# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MathMind AI EdTech Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # Subjects
...
  /subjects/{id}/topics:
    get:
      operationId: listTopics
      tags: [subjects]
      summary: List topics for a subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of topics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Topic"

  # Practice
...
  /progress:
    get:
      operationId: getProgress
      tags: [progress]
      summary: Get student progress overview
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Progress overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressOverview"
  /progress/topic-stats:
    get:
      operationId: getTopicStats
      tags: [progress]
      summary: Get per-topic stats
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Topic stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopicStat"
  /progress/accuracy-history:
    get:
      operationId: getAccuracyHistory
      tags: [progress]
      summary: Get accuracy over time (for graph)
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: days
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Accuracy history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccuracyPoint"

  # Dashboard
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # Auth
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        language:
          type: string
          enum: [en, hi]
        xp:
          type: integer
        level:
          type: integer
        streak:
          type: integer
        totalQuestionsAnswered:
          type: integer
        totalCorrect:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
        - displayName
        - language
        - xp
        - level
        - streak
        - totalQuestionsAnswered
        - totalCorrect
        - createdAt
    UpdateUserBody:
      type: object
      properties:
        displayName:
          type: string
        avatarUrl:
          type: string
        language:
          type: string
          enum: [en, hi]

    # Subjects & Topics
...
    Topic:
      type: object
      properties:
        id:
          type: integer
        subjectId:
          type: integer
        name:
          type: string
        nameHi:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        order:
          type: integer
      required:
        - id
        - subjectId
        - name
        - nameHi
        - difficulty
        - order

    # Practice
...
    ProgressOverview:
      type: object
      properties:
        totalXp:
          type: integer
        level:
          type: integer
        streak:
          type: integer
        totalAnswered:
          type: integer
        totalCorrect:
          type: integer
        overallAccuracy:
          type: number
        studyTimeMinutes:
          type: integer
        weakTopics:
          type: array
          items:
            type: string
        strongTopics:
          type: array
          items:
            type: string
      required:
        - totalXp
        - level
        - streak
        - totalAnswered
        - totalCorrect
        - overallAccuracy
        - studyTimeMinutes
        - weakTopics
        - strongTopics
    TopicStat:
      type: object
      properties:
        topicId:
          type: integer
        topicName:
          type: string
        subjectName:
          type: string
        accuracy:
          type: number
        answered:
          type: integer
        correct:
          type: integer
        strength:
          type: string
          enum: [weak, average, strong]
      required:
        - topicId
        - topicName
        - subjectName
        - accuracy
        - answered
        - correct
        - strength
    AccuracyPoint:
      type: object
      properties:
        date:
          type: string
        accuracy:
          type: number
        questionsAnswered:
          type: integer
      required:
        - date
        - accuracy
        - questionsAnswered

    # Dashboard
...